Changed double position and speed variables to floats to allow for more steering precision

+++ src/org/usfirst/frc/team1571/robot/commands/DriveManual.java
@@ -7,9 +7,9 @@ import edu.wpi.first.wpilibj.command.Command;
 public class DriveManual extends Command {
 	
 	private boolean isFinished;
-	private double steerPosition, speed;
+	private float steerPosition, speed;
 
-    public DriveManual(double steerPosition, double speed) {
+    public DriveManual(float steerPosition, float speed) {
         requires(Robot.driveSystem);
         
         this.steerPosition = steerPosition;
